adicionado testes de repositório, iniciado states

+++ lib/app/models/movie_model.dart
@@ -16,7 +16,7 @@ class MovieModel {
       posterPath: json['poster_path'],
       title: json['title'],
       releaseDate: json['release_date'],
-      voteAverage: json['vote_average'],
+      voteAverage: (json['vote_average']).toDouble(),
     );
   }
 }
